% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encodingRoxygen2Error-package.R
\docType{package}
\name{encodingRoxygen2Error-package}
\alias{encodingRoxygen2Error}
\alias{encodingRoxygen2Error-package}
\title{encodingRoxygen2Error: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Lluís Revilla Sancho \email{lluis.revilla@gmail.com} (\href{https://orcid.org/0000-0001-9747-2570}{ORCID})

}
\keyword{internal}
